Fix clippy warnings, formatting
